added sendemail to code model
