+++ server/api/reward/reward.controller.js
@@ -20,35 +20,34 @@ exports.get = function(req, res) {
 };
 
 exports.redeem = function(req, res) {
-	var redeem = req.body;
+	var body = req.body;
 	Code.forge({
-		reward_id: redeem.reward_id
+		reward_id: body.reward_id
 	})
 	.fetch()
 	.then(function(code){
 		return code.save({used:true}, {patch:true})
 	})
 	.then(function(code){
-		
 		//decrement the wallet
 		return Wallet.forge({
-			user_id: redeem.user_id
+			user_id: body.user_id
 		})
 		.fetch({require: true})
 		.then(function (wallet) {
 			var amount = wallet.get('amount');
-			amount = parseInt(amount) - parseInt(redeem.amount);
+			amount = parseInt(amount) - parseInt(body.amount);
 			wallet.set('amount', amount);
 			return wallet.save();
 		})
-		.then(function(){
-			return Reward.redeem(user, reward);
+		.then(function(user){
+			console.log(Code.sendEmail);
+			return code.sendEmail(body);
 		})
 		.then(function(){
 			return res.json(code);
 		})
 		.catch(function(err){
-			console.error('err',err);
 			res.status(400).send({error:err.message});
 		});
 	})
